fix y coordinate of printed text on the terminal

+++ include/yarrr/terminal.hpp
@@ -22,13 +22,13 @@ class ChatMessage;
 class Terminal : public GraphicalObject
 {
   public:
-    Terminal( GraphicalEngine&, the::ctci::Dispatcher&, int number_of_messages );
+    Terminal( GraphicalEngine&, the::ctci::Dispatcher&, size_t number_of_messages );
     virtual void draw() const override;
 
   private:
     void handle_chat_message( const ChatMessage& );
     std::vector< std::string > m_messages;
-    const int m_number_of_shown_messages;
+    const size_t m_number_of_shown_messages;
 };
 
 }
